backend/requirements.txt
# Core
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.3
pydantic-settings==2.9.1

# Database
sqlalchemy[asyncio]==2.0.41
asyncpg==0.30.0
alembic==1.15.2

# Redis
redis[hiredis]>=5.3.0

# AWS
boto3==1.38.25
strands-agents==0.1.9
strands-agents-tools==0.1.5
a2a-sdk==0.3.26

# Algorand
py-algorand-sdk==2.11.1

# Auth
python-jose[cryptography]==3.4.0
cryptography==45.0.3
PyJWT>=2.10.1
bcrypt>=4.2.0

# HTTP
httpx==0.28.1

# Testing (dev)
pytest==8.4.0
pytest-asyncio==1.3.0
pytest-cov==6.1.1
aiosqlite==0.22.1

# Utilities
python-multipart==0.0.20
python-dotenv==1.1.0
pdfplumber>=0.11.0
